@@ -52,9 +52,11 @@
     "fast-json-stringify": "^1.5.4",
     "fast-safe-stringify": "^1.2.0",
     "fastify": "^1.1.1",
+    "fastify-cors": "^0.2.0",
     "fastify-formbody": "^2.0.0",
     "fastify-multipart": "^0.4.1",
     "grpc": "^1.14.1",
+    "http2": "^3.3.7",
     "iterare": "0.0.8",
     "json-socket": "^0.2.1",
     "mqtt": "^2.16.0",
@@ -114,14 +116,14 @@
     "mocha": "^3.2.0",
     "nodemon": "^1.18.4",
     "nyc": "^10.1.2",
-    "prettier": "^1.9.2",
+    "prettier": "^1.15.3",
     "sinon": "^2.1.0",
     "sinon-chai": "^2.8.0",
     "socket.io-client": "^2.0.4",
     "supertest": "^3.0.0",
-    "ts-node": "^6.0.0",
+    "ts-node": "^7.0.1",
     "tslint": "^5.11.0",
-    "typescript": "^3.1.6"
+    "typescript": "^3.2.2"
   },
   "collective": {
     "type": "opencollective",
@@ -137,8 +139,7 @@
     "exclude": [
       "node_modules/",
       "packages/**/*.spec.ts",
-      "packages/core/adapters/*.ts",
-      "packages/websockets/adapters/*.ts",
+      "packages/**/adapters/*.ts",
       "packages/**/nest-*.ts",
       "packages/core/errors/**/*",
       "packages/common/exceptions/*.ts",
