@@ -1,123 +1,47 @@
-import {
-  INestApplicationContext,
-  Logger,
-  WebSocketAdapter,
-} from '@nestjs/common';
-import { loadPackage } from '@nestjs/common/utils/load-package.util';
+import { INestApplicationContext, WebSocketAdapter } from '@nestjs/common';
+import { WsMessageHandler } from '@nestjs/common/interfaces';
 import { isFunction } from '@nestjs/common/utils/shared.utils';
 import { NestApplication } from '@nestjs/core';
-import { Server } from 'http';
-import { EMPTY as empty, fromEvent, Observable } from 'rxjs';
-import { filter, first, mergeMap, share, takeUntil } from 'rxjs/operators';
-import { CLOSE_EVENT, CONNECTION_EVENT, ERROR_EVENT } from '../constants';
-import { MessageMappingProperties } from '../gateway-metadata-explorer';
-import { BaseWsExceptionFilter } from './../exceptions/base-ws-exception-filter';
+import { Observable } from 'rxjs';
+import { CONNECTION_EVENT, DISCONNECT_EVENT } from '../constants';
 
-let wsPackage: any = {};
-
-enum READY_STATE {
-  CONNECTING_STATE = 0,
-  OPEN_STATE = 1,
-  CLOSING_STATE = 2,
-  CLOSED_STATE = 3,
+export interface BaseWsInstance {
+  on: (event: string, callback: Function) => void;
+  close: Function;
 }
 
-export class WsAdapter implements WebSocketAdapter {
-  protected readonly baseExceptionFilter = new BaseWsExceptionFilter();
-  protected readonly logger = new Logger(WsAdapter.name);
-  protected readonly httpServer: Server;
+export abstract class AbstractWsAdapter<
+  TServer extends BaseWsInstance = any,
+  TClient extends BaseWsInstance = any,
+  TOptions = any
+> implements WebSocketAdapter<TServer, TClient, TOptions> {
+  protected readonly httpServer: any;
 
-  constructor(appOrHttpServer?: INestApplicationContext | Server) {
-    wsPackage = loadPackage('ws', 'WsAdapter');
+  constructor(appOrHttpServer?: INestApplicationContext | any) {
     if (appOrHttpServer && appOrHttpServer instanceof NestApplication) {
       this.httpServer = appOrHttpServer.getUnderlyingHttpServer();
     } else {
-      this.httpServer = appOrHttpServer as Server;
+      this.httpServer = appOrHttpServer;
     }
   }
 
-  public create(
-    port: number,
-    options?: any & { namespace?: string; server?: any },
-  ): any {
-    const { server, ...wsOptions } = options;
-    if (port === 0 && this.httpServer) {
-      return this.bindErrorHandler(
-        new wsPackage.Server({
-          server: this.httpServer,
-          ...wsOptions,
-        }),
-      );
-    }
-    return server
-      ? server
-      : this.bindErrorHandler(
-          new wsPackage.Server({
-            port,
-            ...wsOptions,
-          }),
-        );
-  }
-
-  public bindClientConnect(server, callback: (...args: any[]) => void) {
+  public bindClientConnect(server: TServer, callback: Function) {
     server.on(CONNECTION_EVENT, callback);
   }
 
-  public bindClientDisconnect(client, callback: (...args: any[]) => void) {
-    client.on(CLOSE_EVENT, callback);
+  public bindClientDisconnect(client: TClient, callback: Function) {
+    client.on(DISCONNECT_EVENT, callback);
   }
 
-  public bindMessageHandlers(
-    client: any,
-    handlers: MessageMappingProperties[],
+  public close(server: TServer) {
+    const isCallable = server && isFunction(server.close);
+    isCallable && server.close();
+  }
+
+  public abstract create(port: number, options?: TOptions): TServer;
+  public abstract bindMessageHandlers(
+    client: TClient,
+    handlers: WsMessageHandler[],
     transform: (data: any) => Observable<any>,
-  ) {
-    const close$ = fromEvent(client, CLOSE_EVENT).pipe(share(), first());
-    const source$ = fromEvent(client, 'message').pipe(
-      mergeMap(data =>
-        this.bindMessageHandler(data, handlers, transform).pipe(
-          filter(result => result),
-        ),
-      ),
-      takeUntil(close$),
-    );
-    const onMessage = (response: any) => {
-      if (client.readyState !== READY_STATE.OPEN_STATE) {
-        return;
-      }
-      client.send(JSON.stringify(response));
-    };
-    const onError = (err: any) =>
-      this.baseExceptionFilter.handleError(client, err);
-    source$.subscribe(onMessage, onError);
-  }
-
-  public bindMessageHandler(
-    buffer: any,
-    handlers: MessageMappingProperties[],
-    transform: (data: any) => Observable<any>,
-  ): Observable<any> {
-    try {
-      const message = JSON.parse(buffer.data);
-      const messageHandler = handlers.find(
-        handler => handler.message === message.event,
-      );
-      const { callback } = messageHandler;
-      return transform(callback(message.data));
-    } catch {
-      return empty;
-    }
-  }
-
-  public close(server: any) {
-    server && isFunction(server.close) && server.close();
-  }
-
-  public bindErrorHandler(server: any) {
-    server.on(CONNECTION_EVENT, ws =>
-      ws.on(ERROR_EVENT, (err: any) => this.logger.error(err)),
-    );
-    server.on(ERROR_EVENT, (err: any) => this.logger.error(err));
-    return server;
-  }
+  );
 }
