@@ -12,6 +12,16 @@ const packages = {
   microservices: ts.createProject('packages/microservices/tsconfig.json'),
   websockets: ts.createProject('packages/websockets/tsconfig.json'),
   testing: ts.createProject('packages/testing/tsconfig.json'),
+  'platform-express': ts.createProject(
+    'packages/platform-express/tsconfig.json',
+  ),
+  'platform-fastify': ts.createProject(
+    'packages/platform-fastify/tsconfig.json',
+  ),
+  'platform-socket.io': ts.createProject(
+    'packages/platform-socket.io/tsconfig.json',
+  ),
+  'platform-ws': ts.createProject('packages/platform-ws/tsconfig.json'),
 };
 const modules = Object.keys(packages);
 const source = 'packages';
@@ -34,7 +44,11 @@ gulp.task('copy-misc', function() {
     .pipe(gulp.dest(`${source}/core`))
     .pipe(gulp.dest(`${source}/microservices`))
     .pipe(gulp.dest(`${source}/websockets`))
-    .pipe(gulp.dest(`${source}/testing`));
+    .pipe(gulp.dest(`${source}/testing`))
+    .pipe(gulp.dest(`${source}/platform-fastify`))
+    .pipe(gulp.dest(`${source}/platform-express`))
+    .pipe(gulp.dest(`${source}/platform-ws`))
+    .pipe(gulp.dest(`${source}/platform-socket.io`));
 });
 
 gulp.task('clean:output', function() {
