@@ -1,7 +1,8 @@
+import { INestApplication } from '@nestjs/common';
+import { ExpressAdapter } from '@nestjs/platform-express';
+import { Test } from '@nestjs/testing';
 import * as express from 'express';
 import * as request from 'supertest';
-import { Test } from '@nestjs/testing';
-import { INestApplication } from '@nestjs/common';
 import { ApplicationModule } from '../src/app.module';
 
 describe('Hello world (express instance)', () => {
@@ -11,10 +12,9 @@ describe('Hello world (express instance)', () => {
   beforeEach(async () => {
     const module = await Test.createTestingModule({
       imports: [ApplicationModule],
-    })
-      .compile();
+    }).compile();
 
-    app = module.createNestApplication(express());
+    app = module.createNestApplication(new ExpressAdapter(express()));
     server = app.getHttpServer();
     await app.init();
   });
