@@ -1,19 +1,22 @@
-import { INestApplication } from '@nestjs/common';
-import { INestFastifyApplication } from '@nestjs/common/interfaces/nest-fastify-application.interface';
-import { FastifyAdapter } from '@nestjs/core/adapters/fastify-adapter';
+import {
+  FastifyAdapter,
+  INestFastifyApplication,
+} from '@nestjs/platform-fastify';
 import { Test } from '@nestjs/testing';
 import { expect } from 'chai';
 import { ApplicationModule } from '../src/app.module';
 
 describe('Hello world (fastify adapter)', () => {
-  let app: INestApplication & INestFastifyApplication;
+  let app: INestFastifyApplication;
 
   beforeEach(async () => {
     const module = await Test.createTestingModule({
       imports: [ApplicationModule],
     }).compile();
 
-    app = module.createNestApplication(new FastifyAdapter());
+    app = module.createNestApplication<INestFastifyApplication>(
+      new FastifyAdapter(),
+    );
     await app.init();
   });
 
