@@ -1,24 +1,30 @@
-import * as request from 'supertest';
-import { Test } from '@nestjs/testing';
-import { INestApplication, Injectable } from '@nestjs/common';
-import { ApplicationModule } from '../src/app.module';
+import {
+  CallHandler,
+  ExecutionContext,
+  INestApplication,
+  Injectable,
+  NestInterceptor,
+} from '@nestjs/common';
 import { APP_INTERCEPTOR } from '@nestjs/core';
+import { Test } from '@nestjs/testing';
 import { of } from 'rxjs';
 import { map } from 'rxjs/operators';
+import * as request from 'supertest';
+import { ApplicationModule } from '../src/app.module';
 
 const RETURN_VALUE = 'test';
 
 @Injectable()
-export class OverrideInterceptor {
-  intercept(context, stream) {
+export class OverrideInterceptor implements NestInterceptor {
+  intercept(context: ExecutionContext, next: CallHandler) {
     return of(RETURN_VALUE);
   }
 }
 
 @Injectable()
 export class TransformInterceptor {
-  intercept(context, stream) {
-    return stream.pipe(map(data => ({ data })));
+  intercept(context: ExecutionContext, next: CallHandler) {
+    return next.handle().pipe(map(data => ({ data })));
   }
 }
 
@@ -41,7 +47,7 @@ describe('Interceptors', () => {
     app = (await createTestModule(
       new OverrideInterceptor(),
     )).createNestApplication();
-  
+
     await app.init();
     return request(app.getHttpServer())
       .get('/hello')
@@ -52,7 +58,7 @@ describe('Interceptors', () => {
     app = (await createTestModule(
       new TransformInterceptor(),
     )).createNestApplication();
-  
+
     await app.init();
     return request(app.getHttpServer())
       .get('/hello')
@@ -63,7 +69,7 @@ describe('Interceptors', () => {
     app = (await createTestModule(
       new TransformInterceptor(),
     )).createNestApplication();
-  
+
     await app.init();
     return request(app.getHttpServer())
       .get('/hello/stream')
@@ -74,7 +80,7 @@ describe('Interceptors', () => {
     app = (await createTestModule(
       new TransformInterceptor(),
     )).createNestApplication();
-  
+
     await app.init();
     return request(app.getHttpServer())
       .get('/hello/async')
