@@ -1,8 +1,7 @@
-import * as express from 'express';
-import * as request from 'supertest';
-import { Test } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import { Transport } from '@nestjs/microservices';
+import { Test } from '@nestjs/testing';
+import * as request from 'supertest';
 import { MqttBroadcastController } from '../src/mqtt/mqtt-broadcast.controller';
 
 describe('MQTT transport', () => {
@@ -14,8 +13,9 @@ describe('MQTT transport', () => {
       controllers: [MqttBroadcastController],
     }).compile();
 
-    server = express();
-    app = module.createNestApplication(server);
+    app = module.createNestApplication();
+    server = app.getHttpAdapter().getInstance();
+
     app.connectMicroservice({
       transport: Transport.MQTT,
     });
