@@ -10,7 +10,6 @@ export * from './cache';
 export * from './decorators';
 export * from './enums';
 export * from './exceptions';
-export * from './files';
 export * from './http';
 export {
   ArgumentMetadata,
@@ -24,8 +23,6 @@ export {
   HttpServer,
   INestApplication,
   INestApplicationContext,
-  INestExpressApplication,
-  INestFastifyApplication,
   INestMicroservice,
   MiddlewareConsumer,
   MiddlewareFunction,
@@ -44,8 +41,9 @@ export {
   ValidationError,
   WebSocketAdapter,
   WsExceptionFilter,
+  WsMessageHandler,
 } from './interfaces';
 export * from './pipes';
 export * from './serializer';
-export * from './services/logger.service';
+export * from './services';
 export * from './utils';
