@@ -8,12 +8,11 @@ import 'reflect-metadata';
 
 export * from './adapters';
 export { APP_FILTER, APP_GUARD, APP_INTERCEPTOR, APP_PIPE } from './constants';
-export { BaseExceptionFilter } from './exceptions/base-exception-filter';
-export { ApplicationReferenceHost } from './helpers/application-ref-host';
+export * from './exceptions';
+export * from './helpers';
 export * from './hooks';
-export { ModuleRef } from './injector/module-ref';
-export { HTTP_SERVER_REF } from './injector/tokens';
-export { MiddlewareBuilder } from './middleware/builder';
+export * from './injector';
+export * from './middleware';
 export * from './nest-application';
 export * from './nest-application-context';
 export { NestFactory } from './nest-factory';
