@@ -3,7 +3,7 @@ import { CallHandler, Controller } from '@nestjs/common/interfaces';
 import { isEmpty } from '@nestjs/common/utils/shared.utils';
 import { defer, from as fromPromise, Observable } from 'rxjs';
 import { mergeAll, switchMap } from 'rxjs/operators';
-import { ExecutionContextHost } from '../helpers/execution-context.host';
+import { ExecutionContextHost } from '../helpers/execution-context-host';
 
 export class InterceptorsConsumer {
   public async intercept(
