@@ -1,6 +1,6 @@
 import * as sinon from 'sinon';
 import { expect } from 'chai';
-import { ExecutionContextHost } from '../../helpers/execution-context.host';
+import { ExecutionContextHost } from '../../helpers/execution-context-host';
 
 describe('ExecutionContextHost', () => {
   let contextHost: ExecutionContextHost;
