@@ -6,8 +6,7 @@
  */
 import 'reflect-metadata';
 
-export * from './adapters/io-adapter';
-export * from './adapters/ws-adapter';
+export * from './adapters';
 export * from './errors';
 export * from './exceptions';
 export { MessageMappingProperties } from './gateway-metadata-explorer';
