@@ -1,9 +1,16 @@
-import { NestFactory, FastifyAdapter } from '@nestjs/core';
 import { ValidationPipe } from '@nestjs/common';
+import { NestFactory } from '@nestjs/core';
+import {
+  FastifyAdapter,
+  INestFastifyApplication,
+} from '@nestjs/platform-fastify';
 import { ApplicationModule } from './app.module';
 
 async function bootstrap() {
-  const app = await NestFactory.create(ApplicationModule, new FastifyAdapter());
+  const app = await NestFactory.create<INestFastifyApplication>(
+    ApplicationModule,
+    new FastifyAdapter(),
+  );
   app.useGlobalPipes(new ValidationPipe());
   await app.listen(3000);
 }
